+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# Work in Progress
-* Switch to protobufjs v6 (2.2.0)
+# In progress
+* 
+
+# 2.3.x
+* Switch to protobufjs v6 (2.3.0) - not published
+* 0.67.3 (2.2.0)
 * Improve Google login error handling (2.1.1)
 * Add `client.getOption()` to get an option current value (2.1.1)
 * Minor text fix